wip(cut-c): validate/v1 removal in progress (build red, 4 files remain)
Checkpoint snapshot — NOT green. Detector gate, ghost-core/checks, govern, the
check command, file-kind, init, package-context, package-review-command done;
parseUnifiedDiff rescued to scan/unified-diff.ts. Remaining: scan-status,
fingerprint-stack, verify-package, fingerprint-contribution (validate woven in
as a scan stage / contribution count / verify dimension / stack plumbing).
