--- packages/ghost/src/core/check.ts
@@ -1,452 +1,0 @@
-import { execFile } from "node:child_process";
-import { promisify } from "node:util";
-import { parse as parseYaml } from "yaml";
-import {
-  GHOST_VALIDATE_SCHEMA,
-  type GhostCheck,
-  type GhostValidateDocument,
-  GhostValidateSchema,
-  lintGhostValidate,
-  routeGhostValidateForPath,
-} from "#ghost-core";
-import { readOptionalUtf8 } from "../internal/fs.js";
-import {
-  loadFingerprintPackage,
-  resolveFingerprintPackage,
-} from "../scan/fingerprint-package.js";
-import {
-  groupFingerprintStacksForPaths,
-  resolveGhostDirDefault,
-} from "../scan/fingerprint-stack.js";
-import {
-  INLINE_COLOR_LITERAL_PATTERN,
-  isInlineColorDetector,
-} from "./inline-color-literals.js";
-
-const execFileAsync = promisify(execFile);
-
-export interface GhostDriftCheckOptions {
-  cwd?: string;
-  packageDir?: string;
-  ghostDir?: string;
-  base?: string;
-  diffText?: string;
-}
-
-export interface GhostDriftChangedLine {
-  path: string;
-  line: number;
-  text: string;
-}
-
-export interface GhostDriftChangedFile {
-  path: string;
-  added_lines: GhostDriftChangedLine[];
-}
-
-export interface GhostDriftRoutedFile {
-  path: string;
-  scopes: string[];
-  checks: string[];
-}
-
-export interface GhostDriftCheckFinding {
-  check_id: string;
-  title: string;
-  severity: GhostCheck["severity"];
-  path: string;
-  line: number;
-  detector: GhostCheck["detector"]["type"];
-  message: string;
-  repair?: string;
-  match?: string;
-}
-
-export interface GhostDriftCheckReport {
-  schema: "ghost.check-report/v1";
-  result: "pass" | "fail";
-  package_dir: string;
-  ghost_dir?: string;
-  base?: string;
-  changed_files: string[];
-  routed_files: GhostDriftRoutedFile[];
-  findings: GhostDriftCheckFinding[];
-  stacks?: GhostDriftCheckStack[];
-}
-
-export interface GhostDriftCheckStack {
-  target_path: string;
-  package_dir: string;
-  ghost_dir: string;
-  changed_files: string[];
-  stack_dirs: string[];
-  provenance: {
-    stack: Array<{
-      dir: string;
-      root: string;
-      relative_root: string;
-    }>;
-  };
-}
-
-interface LoadedCheckPackage {
-  dir: string;
-  checks: GhostValidateDocument;
-}
-
-export async function runGhostDriftCheck(
-  options: GhostDriftCheckOptions = {},
-): Promise<GhostDriftCheckReport> {
-  const cwd = options.cwd ?? process.cwd();
-  const diffText =
-    options.diffText ??
-    (await readGitDiff(cwd, options.base ?? "HEAD", "--unified=0"));
-  const changedFiles = parseUnifiedDiff(diffText);
-
-  if (options.packageDir) {
-    const pkg = await loadCheckPackage(options.packageDir, cwd);
-    const evaluated = evaluateChangedFiles(changedFiles, pkg);
-    return {
-      schema: "ghost.check-report/v1",
-      result: evaluated.findings.length > 0 ? "fail" : "pass",
-      package_dir: pkg.dir,
-      ...(options.base ? { base: options.base } : {}),
-      changed_files: changedFiles.map((file) => file.path),
-      routed_files: evaluated.routedFiles,
-      findings: evaluated.findings,
-    };
-  }
-
-  const groups = await groupFingerprintStacksForPaths(
-    changedFiles.map((file) => file.path),
-    cwd,
-    { ghostDir: resolveGhostDirDefault(options.ghostDir) },
-  );
-  const routedFiles: GhostDriftRoutedFile[] = [];
-  const findings: GhostDriftCheckFinding[] = [];
-  const stacks: GhostDriftCheckStack[] = [];
-
-  for (const group of groups) {
-    const filesForStack = changedFiles.filter((file) =>
-      group.changed_files.includes(file.path),
-    );
-    const leaf = group.stack.layers.at(-1);
-    const pkg: LoadedCheckPackage = {
-      dir: leaf?.dir ?? group.stack.layers[0].dir,
-      checks: group.stack.contract.checks,
-    };
-    const evaluated = evaluateChangedFiles(filesForStack, pkg);
-    routedFiles.push(...evaluated.routedFiles);
-    findings.push(...evaluated.findings);
-    stacks.push({
-      target_path: group.stack.target_path,
-      package_dir: pkg.dir,
-      ghost_dir: group.stack.ghost_dir,
-      changed_files: group.changed_files,
-      stack_dirs: group.stack.layers.map((layer) => layer.dir),
-      provenance: {
-        stack: group.stack.provenance.layers,
-      },
-    });
-  }
-
-  return {
-    schema: "ghost.check-report/v1",
-    result: findings.length > 0 ? "fail" : "pass",
-    package_dir:
-      stacks.length === 1
-        ? stacks[0].package_dir
-        : "fingerprint-stack/multiple",
-    ghost_dir: stacks[0]?.ghost_dir ?? options.ghostDir,
-    ...(options.base ? { base: options.base } : {}),
-    changed_files: changedFiles.map((file) => file.path),
-    routed_files: routedFiles,
-    findings,
-    stacks,
-  };
-}
-
-export function parseUnifiedDiff(diffText: string): GhostDriftChangedFile[] {
-  const files = new Map<string, GhostDriftChangedFile>();
-  let current: GhostDriftChangedFile | undefined;
-  let newLine = 0;
-
-  for (const rawLine of diffText.split(/\r?\n/)) {
-    if (rawLine.startsWith("diff --git ")) {
-      current = undefined;
-      continue;
-    }
-
-    if (rawLine.startsWith("+++ ")) {
-      const file = rawLine.replace(/^\+\+\+\s+/, "");
-      if (file === "/dev/null") {
-        current = undefined;
-        continue;
-      }
-      const path = file.replace(/^b\//, "");
-      current = files.get(path) ?? { path, added_lines: [] };
-      files.set(path, current);
-      continue;
-    }
-
-    const hunk = rawLine.match(/^@@\s+-\d+(?:,\d+)?\s+\+(\d+)(?:,\d+)?\s+@@/);
-    if (hunk) {
-      newLine = Number(hunk[1]);
-      continue;
-    }
-
-    if (!current) continue;
-    if (rawLine.startsWith("+")) {
-      current.added_lines.push({
-        path: current.path,
-        line: newLine,
-        text: rawLine.slice(1),
-      });
-      newLine += 1;
-    } else if (rawLine.startsWith("-")) {
-    } else {
-      newLine += 1;
-    }
-  }
-
-  return [...files.values()];
-}
-
-export function formatGhostDriftCheckMarkdown(
-  report: GhostDriftCheckReport,
-): string {
-  const lines = [
-    `Design Check: ${report.result === "pass" ? "PASS" : "FAIL"}`,
-    `Package: ${report.package_dir}`,
-    `Changed files: ${report.changed_files.length}`,
-    `Findings: ${report.findings.length}`,
-    "",
-  ];
-
-  if (report.routed_files.length > 0) {
-    lines.push("## Routed Files", "");
-    for (const file of report.routed_files) {
-      const scopes = file.scopes.length > 0 ? file.scopes.join(", ") : "none";
-      const checks = file.checks.length > 0 ? file.checks.join(", ") : "none";
-      lines.push(`- ${file.path}: scopes ${scopes}; checks ${checks}`);
-    }
-    lines.push("");
-  }
-
-  if (report.findings.length === 0) {
-    lines.push("No active deterministic check failures.");
-    return `${lines.join("\n")}\n`;
-  }
-
-  lines.push("## Issues", "");
-  report.findings.forEach((finding, index) => {
-    lines.push(
-      `${index + 1}. [${finding.severity}] ${finding.title} (${finding.check_id})`,
-      `   ${finding.path}:${finding.line} — ${finding.message}`,
-    );
-    if (finding.match) lines.push(`   Match: \`${finding.match}\``);
-    if (finding.repair) lines.push(`   Repair: ${finding.repair}`);
-  });
-  return `${lines.join("\n")}\n`;
-}
-
-async function loadCheckPackage(
-  packageDir: string | undefined,
-  cwd: string,
-): Promise<LoadedCheckPackage> {
-  const paths = resolveFingerprintPackage(packageDir, cwd);
-  const [loaded, checksRaw] = await Promise.all([
-    loadFingerprintPackage(paths),
-    readOptional(paths.checks),
-  ]);
-  const fingerprint = loaded.fingerprint;
-  if (checksRaw === undefined) {
-    return {
-      dir: paths.dir,
-      checks: {
-        schema: GHOST_VALIDATE_SCHEMA,
-        id: "none",
-        checks: [],
-      },
-    };
-  }
-  const checksInput = parseYaml(checksRaw);
-  const checksResult = GhostValidateSchema.safeParse(checksInput);
-  if (!checksResult.success) {
-    throw new Error(
-      `validate.yml failed schema validation: ${checksResult.error.issues
-        .map((issue) => `${issue.path.join(".") || "<root>"}: ${issue.message}`)
-        .join("; ")}`,
-    );
-  }
-  const checks = checksResult.data as GhostValidateDocument;
-  const checkLint = lintGhostValidate(checks, { fingerprint });
-  if (checkLint.errors > 0) {
-    throw new Error(
-      `validate.yml failed lint with ${checkLint.errors} error(s): ${checkLint.issues
-        .filter((issue) => issue.severity === "error")
-        .map((issue) => `[${issue.rule}] ${issue.message}`)
-        .join("; ")}`,
-    );
-  }
-  return { dir: paths.dir, checks };
-}
-
-function evaluateChangedFiles(
-  changedFiles: GhostDriftChangedFile[],
-  pkg: LoadedCheckPackage,
-): {
-  routedFiles: GhostDriftRoutedFile[];
-  findings: GhostDriftCheckFinding[];
-} {
-  const routedFiles: GhostDriftRoutedFile[] = [];
-  const findings: GhostDriftCheckFinding[] = [];
-
-  for (const file of changedFiles) {
-    const routed = routeGhostValidateForPath(pkg.checks.checks, file.path);
-    routedFiles.push({
-      path: file.path,
-      scopes: [],
-      checks: routed.map((entry) => entry.check.id),
-    });
-
-    for (const entry of routed) {
-      if (!detectorAppliesToPath(entry.check, file.path)) continue;
-      findings.push(...evaluateCheck(entry.check, file));
-    }
-  }
-
-  return { routedFiles, findings };
-}
-
-const readOptional = readOptionalUtf8;
-
-async function readGitDiff(
-  cwd: string,
-  base: string,
-  contextFlag: string,
-): Promise<string> {
-  const { stdout } = await execFileAsync("git", ["diff", contextFlag, base], {
-    cwd,
-    maxBuffer: 1024 * 1024 * 20,
-  });
-  return stdout;
-}
-
-function evaluateCheck(
-  check: GhostCheck,
-  file: GhostDriftChangedFile,
-): GhostDriftCheckFinding[] {
-  const regexes = detectorRegexes(check);
-  if (regexes.length === 0) return [];
-
-  if (isRequiredDetector(check)) {
-    if (file.added_lines.length === 0) return [];
-    const hasMatch = file.added_lines.some((line) => {
-      return regexes.some((regex) => {
-        regex.lastIndex = 0;
-        return regex.test(line.text);
-      });
-    });
-    if (hasMatch) return [];
-    const firstLine = file.added_lines[0];
-    return [
-      {
-        check_id: check.id,
-        title: check.title,
-        severity: check.severity,
-        path: file.path,
-        line: firstLine.line,
-        detector: check.detector.type,
-        message: requiredMessage(check),
-        ...(check.repair ? { repair: check.repair } : {}),
-      },
-    ];
-  }
-
-  const findings: GhostDriftCheckFinding[] = [];
-  const seen = new Set<string>();
-  for (const line of file.added_lines) {
-    for (const regex of regexes) {
-      regex.lastIndex = 0;
-      let match = regex.exec(line.text);
-      while (match !== null) {
-        const key = `${line.line}:${match.index}:${match[0]}`;
-        if (!seen.has(key)) {
-          findings.push({
-            check_id: check.id,
-            title: check.title,
-            severity: check.severity,
-            path: file.path,
-            line: line.line,
-            detector: check.detector.type,
-            message: forbiddenMessage(check),
-            match: match[0],
-            ...(check.repair ? { repair: check.repair } : {}),
-          });
-          seen.add(key);
-        }
-        if (match[0] === "") regex.lastIndex += 1;
-        match = regex.exec(line.text);
-      }
-    }
-  }
-  return findings;
-}
-
-function detectorRegexes(check: GhostCheck): RegExp[] {
-  const source =
-    check.detector.pattern ??
-    (check.detector.value ? escapeRegExp(check.detector.value) : undefined);
-  if (!source) return [];
-
-  const regexes = [new RegExp(source, "g")];
-  if (isInlineColorDetector(check, source)) {
-    regexes.push(new RegExp(INLINE_COLOR_LITERAL_PATTERN, "gi"));
-  }
-  return regexes;
-}
-
-function detectorAppliesToPath(check: GhostCheck, path: string): boolean {
-  const contexts = check.detector.contexts;
-  if (!contexts?.length) return true;
-  const context = contextForPath(path);
-  return context ? contexts.includes(context) : false;
-}
-
-function contextForPath(path: string): string | undefined {
-  const lower = path.toLowerCase();
-  if (lower.endsWith(".swift")) return "swift";
-  if (lower.endsWith(".tsx") || lower.endsWith(".jsx")) return "react";
-  if (lower.endsWith(".ts") || lower.endsWith(".js")) return "typescript";
-  if (lower.endsWith(".css") || lower.endsWith(".scss")) return "css";
-  return undefined;
-}
-
-function isRequiredDetector(check: GhostCheck): boolean {
-  return (
-    check.detector.type === "required-regex" ||
-    check.detector.type === "required-token"
-  );
-}
-
-function requiredMessage(check: GhostCheck): string {
-  if (check.detector.type === "required-token") {
-    return "Added UI code did not use the required design token.";
-  }
-  return "Added UI code did not match the required pattern.";
-}
-
-function forbiddenMessage(check: GhostCheck): string {
-  if (check.detector.type === "banned-import") {
-    return "Added UI code imports a banned dependency.";
-  }
-  if (check.detector.type === "banned-component") {
-    return "Added UI code uses a banned component.";
-  }
-  return "Added UI code matched a forbidden pattern.";
-}
-
-function escapeRegExp(value: string): string {
-  return value.replace(/[|\\{}()[\]^$+?.]/g, "\\$&");
-}
