--- packages/ghost/src/ghost-core/checks/index.ts
@@ -1,34 +1,0 @@
-export { lintGhostValidate } from "./lint.js";
-export {
-  matchesGhostPath,
-  normalizeGhostPath,
-  routeGhostValidateForPath,
-} from "./routing.js";
-export {
-  GhostCheckDerivationSchema,
-  GhostCheckSchema,
-  GhostValidateSchema,
-} from "./schema.js";
-export type {
-  GhostCheck,
-  GhostCheckAppliesTo,
-  GhostCheckDerivation,
-  GhostCheckDerivationCompositionRef,
-  GhostCheckDerivationIntentRef,
-  GhostCheckDerivationInventoryRef,
-  GhostCheckDetector,
-  GhostCheckDetectorType,
-  GhostCheckEvidence,
-  GhostCheckSeverity,
-  GhostCheckStatus,
-  GhostValidateDocument,
-  GhostValidateLintIssue,
-  GhostValidateLintOptions,
-  GhostValidateLintReport,
-  GhostValidateLintSeverity,
-  RoutedGhostValidateCheck,
-} from "./types.js";
-export {
-  GHOST_VALIDATE_FILENAME,
-  GHOST_VALIDATE_SCHEMA,
-} from "./types.js";
