--- packages/ghost/src/ghost-core/checks/types.ts
@@ -1,99 +1,0 @@
-import type {
-  GhostFingerprintDocument,
-  GhostFingerprintRef,
-} from "../fingerprint/index.js";
-
-export const GHOST_VALIDATE_SCHEMA = "ghost.validate/v1" as const;
-export const GHOST_VALIDATE_FILENAME = "validate.yml" as const;
-
-export type GhostCheckStatus = "active" | "proposed" | "disabled";
-export type GhostCheckSeverity = "critical" | "serious" | "nit";
-export type GhostCheckDerivationIntentRef = Extract<
-  GhostFingerprintRef,
-  | `intent.principle:${string}`
-  | `intent.situation:${string}`
-  | `intent.experience_contract:${string}`
->;
-export type GhostCheckDerivationInventoryRef = Extract<
-  GhostFingerprintRef,
-  `inventory.exemplar:${string}`
->;
-export type GhostCheckDerivationCompositionRef = Extract<
-  GhostFingerprintRef,
-  `composition.pattern:${string}`
->;
-
-export interface GhostCheckDerivation {
-  intent?: GhostCheckDerivationIntentRef[];
-  inventory?: GhostCheckDerivationInventoryRef[];
-  composition?: GhostCheckDerivationCompositionRef[];
-}
-
-export type GhostCheckDetectorType =
-  | "forbidden-regex"
-  | "required-regex"
-  | "banned-import"
-  | "banned-component"
-  | "required-token";
-
-export interface GhostCheckAppliesTo {
-  scopes?: string[];
-  paths?: string[];
-  surface_types?: string[];
-  pattern_ids?: string[];
-}
-
-export interface GhostCheckDetector {
-  type: GhostCheckDetectorType;
-  pattern?: string;
-  value?: string;
-  contexts?: string[];
-}
-
-export interface GhostCheckEvidence {
-  support?: number;
-  observed_count?: number;
-  examples?: Array<string | { path: string; note?: string }>;
-}
-
-export interface GhostCheck {
-  id: string;
-  title: string;
-  status: GhostCheckStatus;
-  severity: GhostCheckSeverity;
-  derivation?: GhostCheckDerivation;
-  applies_to?: GhostCheckAppliesTo;
-  detector: GhostCheckDetector;
-  evidence?: GhostCheckEvidence;
-  repair?: string;
-}
-
-export interface GhostValidateDocument {
-  schema: typeof GHOST_VALIDATE_SCHEMA;
-  id: string;
-  checks: GhostCheck[];
-}
-
-export type GhostValidateLintSeverity = "error" | "warning" | "info";
-
-export interface GhostValidateLintIssue {
-  severity: GhostValidateLintSeverity;
-  rule: string;
-  message: string;
-  path?: string;
-}
-
-export interface GhostValidateLintReport {
-  issues: GhostValidateLintIssue[];
-  errors: number;
-  warnings: number;
-  info: number;
-}
-
-export interface GhostValidateLintOptions {
-  fingerprint?: GhostFingerprintDocument;
-}
-
-export interface RoutedGhostValidateCheck {
-  check: GhostCheck;
-}
