--- packages/ghost/src/govern.ts
@@ -1,14 +1,0 @@
-export type {
-  GhostDriftChangedFile as GhostCheckChangedFile,
-  GhostDriftChangedLine as GhostCheckChangedLine,
-  GhostDriftCheckFinding as GhostCheckFinding,
-  GhostDriftCheckOptions as GhostCheckOptions,
-  GhostDriftCheckReport as GhostCheckReport,
-  GhostDriftCheckStack as GhostCheckStack,
-  GhostDriftRoutedFile as GhostCheckRoutedFile,
-} from "./core/index.js";
-export * from "./core/index.js";
-export {
-  formatGhostDriftCheckMarkdown as formatGhostCheckMarkdown,
-  runGhostDriftCheck as runGhostCheck,
-} from "./core/index.js";
