+++ packages/ghost/src/index.ts
@@ -8,6 +8,5 @@ export const compare = Object.assign(compareFunction, compareApi);
 export * as driftCommand from "./drift-command.js";
 export * as fingerprint from "./fingerprint.js";
 export * as core from "./ghost-core/index.js";
-export * as govern from "./govern.js";
 /** @deprecated Use `fingerprint` or `@anarchitecture/ghost/fingerprint`. */
 export * as scan from "./scan/index.js";
