+++ packages/ghost/src/review-packet.ts
@@ -5,7 +5,7 @@ import {
   type SurfaceGrounding,
   selectChecksForSurfaces,
 } from "#ghost-core";
-import { parseUnifiedDiff } from "./core/index.js";
+import { parseUnifiedDiff } from "./scan/unified-diff.js";
 import { discoverBindingsForPath } from "./scan/binding-discovery.js";
 import { loadChecksDir } from "./scan/checks-dir.js";
 import {
