+++ packages/ghost/src/scan/fingerprint-package.ts
@@ -3,11 +3,9 @@ import { join, resolve } from "node:path";
 import { parse as parseYaml } from "yaml";
 import {
   GHOST_SURFACES_YML_FILENAME,
-  GHOST_VALIDATE_FILENAME,
   type GhostFingerprintDocument,
   type GhostFingerprintPackageManifest,
   type GhostSurfacesDocument,
-  lintGhostValidate,
   SURVEY_FILENAME,
 } from "#ghost-core";
 import {
@@ -53,7 +51,6 @@ export interface FingerprintPackagePaths {
   patterns: string;
   /** Legacy direct markdown path; not part of the canonical root bundle. */
   fingerprint: string;
-  checks: string;
 }
 
 export interface LoadedFingerprintPackage {
@@ -93,7 +90,6 @@ export function resolveFingerprintPackage(
     survey: join(dir, SURVEY_FILENAME),
     patterns: join(dir, PATTERNS_FILENAME),
     fingerprint: join(dir, FINGERPRINT_FILENAME),
-    checks: join(packageDir, GHOST_VALIDATE_FILENAME),
   };
 }
 
@@ -109,7 +105,6 @@ export async function initFingerprintPackage(
     { path: paths.intent, content: templateIntent() },
     { path: paths.inventory, content: templateInventory(options.reference) },
     { path: paths.composition, content: templateComposition() },
-    { path: paths.checks, content: templateChecks() },
   ];
   if (!options.force) {
     await assertInitDoesNotOverwrite(files.map((file) => file.path));
@@ -174,7 +169,6 @@ export async function lintFingerprintPackage(
   const intentRaw = await readOptional(paths.intent);
   const inventoryRaw = await readOptional(paths.inventory);
   const compositionRaw = await readOptional(paths.composition);
-  const checksRaw = await readOptional(paths.checks);
 
   let fingerprint: GhostFingerprintDocument | undefined;
   if (manifestRaw !== undefined) {
@@ -183,14 +177,6 @@ export async function lintFingerprintPackage(
       { intentRaw, inventoryRaw, compositionRaw },
       issues,
     );
-  }
-
-  if (checksRaw !== undefined) {
-    const checks = parseYamlSafe(checksRaw, "validate.yml", issues);
-    if (checks !== undefined) {
-      const checksReport = lintGhostValidate(checks, { fingerprint });
-      issues.push(...prefixIssues("validate.yml", checksReport.issues));
-    }
   }
 
   return finalize(issues);
